Split out exceptions into errors.py
This is a first step for moving out bits of telescope_state.py that are
independent of whether a synchronous or asynchronous telstate is in use.
